add determine_marker() and update delimited
added new function determine_marker() to ioUtils.jl, which is used to determine the marker type, and therefore the marker Type when importing files as PopData. It will parse the first 10 samples and guesstimate the marker depending on the highest allele value of all the genotypes across all loci for those first 10 individuals. Uses all individuals if the infile has less than 10 samples in it.  Also, delimited() is now updated to work with the new PopData structure, and detect_marker was integrated into genepop()